Moved voxel related structs into its own file, removed unnecessary complexity of having seperate compute and direct command queue by having a single command queue abstraction
