+++ src/main.cpp
@@ -3,6 +3,7 @@
 #include "voxel-engine/renderer.hpp"
 #include "voxel-engine/shader_compiler.hpp"
 #include "voxel-engine/timer.hpp"
+#include "voxel-engine/voxel.hpp"
 #include "voxel-engine/window.hpp"
 
 #include "shaders/interop/render_resources.hlsli"
@@ -10,351 +11,6 @@
 #include "imgui.h"
 #include "imgui_impl_dx12.h"
 #include "imgui_impl_win32.h"
-
-// Helper functions to go from 1d to 3d and vice versa.
-static inline size_t convert_to_1d(const DirectX::XMUINT3 index_3d, const size_t N)
-{
-    return index_3d.x + N * (index_3d.y + index_3d.z * N);
-}
-
-static inline DirectX::XMUINT3 convert_to_3d(const size_t index, const size_t N)
-{
-    // For reference, index = x + y * N + z * N * N.
-    const u32 z = static_cast<u32>(index / (N * N));
-    const u32 index_2d = static_cast<u32>(index - z * N * N);
-    const u32 y = static_cast<u32>(index_2d / N);
-    const u32 x = static_cast<u32>(index_2d % N);
-
-    return {x, y, z};
-}
-
-// A voxel is just a value on a regular 3D grid. Think of it as the corners where the cells meet in a 3d grid.
-// For 3d visualization of voxels, A cube is rendered for each voxel where the front lower left corner is the 'voxel
-// position' and has a edge length as specified in the class below.
-struct Voxel
-{
-    static constexpr float EDGE_LENGTH{0.1f};
-    bool m_active{true};
-};
-
-struct Chunk
-{
-    explicit Chunk()
-    {
-        // printf("Constructor of chunk!");
-        m_voxels = new Voxel[NUMBER_OF_VOXELS];
-    }
-
-    Chunk(const Chunk &other) = delete;
-    Chunk &operator=(Chunk &other) = delete;
-
-    Chunk(Chunk &&other) noexcept : m_voxels(std::move(other.m_voxels)), m_chunk_index(other.m_chunk_index)
-    {
-        other.m_voxels = nullptr;
-    }
-
-    Chunk &operator=(Chunk &&other) noexcept
-    {
-        this->m_voxels = std::move(other.m_voxels);
-        this->m_chunk_index = other.m_chunk_index;
-        other.m_voxels = nullptr;
-
-        return *this;
-    }
-
-    ~Chunk()
-    {
-        if (m_voxels)
-        {
-
-            printf("Destructor of chunk!");
-            delete[] m_voxels;
-        }
-    }
-
-    static constexpr u32 NUMBER_OF_VOXELS_PER_DIMENSION = 9u;
-    static constexpr size_t NUMBER_OF_VOXELS =
-        NUMBER_OF_VOXELS_PER_DIMENSION * NUMBER_OF_VOXELS_PER_DIMENSION * NUMBER_OF_VOXELS_PER_DIMENSION;
-
-    // A flattened 3d array of Voxels.
-    Voxel *m_voxels{};
-    size_t m_chunk_index{};
-};
-
-// A class that contains a collection of chunks with data for each stored in hash maps for quick access.
-// The states a chunk can be in:
-// (i) Loaded -> Ready to be rendered.
-// (ii) Setup -> Chunk mesh is ready, but associated buffers may or maynot be ready. Once the buffers are ready, these
-// chunks are moved into the loaded chunks list.
-// (iii) Unloaded -> Buffers are ready, but chunk is not rendered.
-
-struct ChunkManager
-{
-    struct SetupChunkData
-    {
-        Chunk m_chunk{};
-
-        StructuredBuffer m_chunk_position_buffer{};
-        StructuredBuffer m_chunk_color_buffer{};
-
-        // NOTE : This is to cleared once the chunk data is present on the GPU.
-        std::vector<DirectX::XMFLOAT3> m_chunk_position_data{};
-        DirectX::XMFLOAT3 m_chunk_color_data{};
-    };
-
-  public:
-    SetupChunkData _create_chunk(Renderer &renderer, const size_t index)
-    {
-        SetupChunkData setup_chunk_data{};
-
-        // Iterate over each voxel in chunk and setup the chunk common position buffer.
-        std::vector<DirectX::XMFLOAT3> position_data{};
-
-        static constexpr std::array<DirectX::XMFLOAT3, 8> voxel_vertices_data{
-            DirectX::XMFLOAT3(0.0f, 0.0f, 0.0f),
-            DirectX::XMFLOAT3(0.0f, Voxel::EDGE_LENGTH, 0.0f),
-            DirectX::XMFLOAT3(Voxel::EDGE_LENGTH, Voxel::EDGE_LENGTH, 0.0f),
-            DirectX::XMFLOAT3(Voxel::EDGE_LENGTH, 0.0f, 0.0f),
-            DirectX::XMFLOAT3(0.0f, 0.0f, Voxel::EDGE_LENGTH),
-            DirectX::XMFLOAT3(0.0f, Voxel::EDGE_LENGTH, Voxel::EDGE_LENGTH),
-            DirectX::XMFLOAT3(Voxel::EDGE_LENGTH, Voxel::EDGE_LENGTH, Voxel::EDGE_LENGTH),
-            DirectX::XMFLOAT3(Voxel::EDGE_LENGTH, 0.0f, Voxel::EDGE_LENGTH),
-        };
-
-        const DirectX::XMUINT3 chunk_index_3d = convert_to_3d(index, ChunkManager::NUMBER_OF_CHUNKS_PER_DIMENSION);
-        const DirectX::XMFLOAT3 chunk_offset =
-            DirectX::XMFLOAT3(chunk_index_3d.x * Voxel::EDGE_LENGTH * Chunk::NUMBER_OF_VOXELS_PER_DIMENSION,
-                              chunk_index_3d.y * Voxel::EDGE_LENGTH * Chunk::NUMBER_OF_VOXELS_PER_DIMENSION,
-                              chunk_index_3d.z * Voxel::EDGE_LENGTH * Chunk::NUMBER_OF_VOXELS_PER_DIMENSION);
-
-        for (size_t i = 0; i < Chunk::NUMBER_OF_VOXELS; i++)
-        {
-            const DirectX::XMUINT3 index_3d = convert_to_3d(i, Chunk::NUMBER_OF_VOXELS_PER_DIMENSION);
-            const DirectX::XMFLOAT3 offset = DirectX::XMFLOAT3(index_3d.x * Voxel::EDGE_LENGTH + chunk_offset.x,
-                                                               index_3d.y * Voxel::EDGE_LENGTH + chunk_offset.y,
-                                                               index_3d.z * Voxel::EDGE_LENGTH + chunk_offset.z);
-
-            // Check if there is a voxel that blocks the front face of current voxel.
-            {
-
-                const bool is_front_face_covered =
-                    (index_3d.z != 0 && setup_chunk_data.m_chunk
-                                            .m_voxels[convert_to_1d({index_3d.x, index_3d.y, index_3d.z - 1},
-                                                                    Chunk::NUMBER_OF_VOXELS_PER_DIMENSION)]
-                                            .m_active);
-
-                if (!is_front_face_covered)
-                {
-                    for (const auto &vertex_index : {0u, 1u, 2u, 0u, 2u, 3u})
-                    {
-                        const auto &vertex = voxel_vertices_data[vertex_index];
-                        position_data.emplace_back(
-                            DirectX::XMFLOAT3{vertex.x + offset.x, vertex.y + offset.y, vertex.z + offset.z});
-                    }
-                }
-            }
-
-            // Check if there is a voxel that blocks the back face of current voxel.
-            {
-
-                const bool is_back_face_covered = (index_3d.z != Chunk::NUMBER_OF_VOXELS_PER_DIMENSION - 1u &&
-                                                   setup_chunk_data.m_chunk
-                                                       .m_voxels[convert_to_1d({index_3d.x, index_3d.y, index_3d.z + 1},
-                                                                               Chunk::NUMBER_OF_VOXELS_PER_DIMENSION)]
-                                                       .m_active);
-
-                if (!is_back_face_covered)
-                {
-                    for (const auto &vertex_index : {4u, 6u, 5u, 4u, 7u, 6u})
-                    {
-                        const auto &vertex = voxel_vertices_data[vertex_index];
-                        position_data.emplace_back(
-                            DirectX::XMFLOAT3{vertex.x + offset.x, vertex.y + offset.y, vertex.z + offset.z});
-                    }
-                }
-            }
-
-            // Check if there is a voxel that blocks the left hand side face of current voxel.
-            {
-
-                const bool is_left_face_covered =
-                    (index_3d.x != 0u && setup_chunk_data.m_chunk
-                                             .m_voxels[convert_to_1d({index_3d.x - 1, index_3d.y, index_3d.z},
-                                                                     Chunk::NUMBER_OF_VOXELS_PER_DIMENSION)]
-                                             .m_active);
-
-                if (!is_left_face_covered)
-                {
-                    for (const auto &vertex_index : {4u, 5u, 1u, 4u, 1u, 0u})
-                    {
-                        const auto &vertex = voxel_vertices_data[vertex_index];
-                        position_data.emplace_back(
-                            DirectX::XMFLOAT3{vertex.x + offset.x, vertex.y + offset.y, vertex.z + offset.z});
-                    }
-                }
-            }
-
-            // Check if there is a voxel that blocks the right hand side face of current voxel.
-            {
-
-                const bool is_right_face_covered =
-                    (index_3d.x != Chunk::NUMBER_OF_VOXELS_PER_DIMENSION - 1u &&
-                     setup_chunk_data.m_chunk
-                         .m_voxels[convert_to_1d({index_3d.x + 1, index_3d.y, index_3d.z},
-                                                 Chunk::NUMBER_OF_VOXELS_PER_DIMENSION)]
-                         .m_active);
-
-                if (!is_right_face_covered)
-                {
-                    for (const auto &vertex_index : {3u, 2u, 6u, 3u, 6u, 7u})
-                    {
-                        const auto &vertex = voxel_vertices_data[vertex_index];
-                        position_data.emplace_back(
-                            DirectX::XMFLOAT3{vertex.x + offset.x, vertex.y + offset.y, vertex.z + offset.z});
-                    }
-                }
-            }
-
-            // Check if there is a voxel that blocks the top side face of current voxel.
-            {
-
-                const bool is_top_face_covered = (index_3d.y != Chunk::NUMBER_OF_VOXELS_PER_DIMENSION - 1 &&
-                                                  setup_chunk_data.m_chunk
-                                                      .m_voxels[convert_to_1d({index_3d.x, index_3d.y + 1, index_3d.z},
-                                                                              Chunk::NUMBER_OF_VOXELS_PER_DIMENSION)]
-                                                      .m_active);
-
-                if (!is_top_face_covered)
-                {
-                    for (const auto &vertex_index : {1u, 5u, 6u, 1u, 6u, 2u})
-                    {
-                        const auto &vertex = voxel_vertices_data[vertex_index];
-                        position_data.emplace_back(
-                            DirectX::XMFLOAT3{vertex.x + offset.x, vertex.y + offset.y, vertex.z + offset.z});
-                    }
-                }
-            }
-
-            // Check if there is a voxel that blocks the bottom side face of current voxel.
-            {
-
-                const bool is_bottom_face_covered =
-                    (index_3d.y != 0u && setup_chunk_data.m_chunk
-                                             .m_voxels[convert_to_1d({index_3d.x, index_3d.y - 1, index_3d.z},
-                                                                     Chunk::NUMBER_OF_VOXELS_PER_DIMENSION)]
-                                             .m_active);
-
-                if (!is_bottom_face_covered)
-                {
-                    for (const auto &vertex_index : {4u, 0u, 3u, 4u, 3u, 7u})
-                    {
-                        const auto &vertex = voxel_vertices_data[vertex_index];
-                        position_data.emplace_back(
-                            DirectX::XMFLOAT3{vertex.x + offset.x, vertex.y + offset.y, vertex.z + offset.z});
-                    }
-                }
-            }
-        }
-
-        setup_chunk_data.m_chunk_position_data = std::move(position_data);
-        setup_chunk_data.m_chunk_color_data = {
-            rand() / (float)RAND_MAX,
-            rand() / (float)RAND_MAX,
-            rand() / (float)RAND_MAX,
-        };
-
-        if (!setup_chunk_data.m_chunk_position_data.empty())
-        {
-            setup_chunk_data.m_chunk_position_buffer = renderer.create_structured_buffer(
-                (void *)setup_chunk_data.m_chunk_position_data.data(), sizeof(DirectX::XMFLOAT3),
-                setup_chunk_data.m_chunk_position_data.size());
-            setup_chunk_data.m_chunk_color_buffer = renderer.create_structured_buffer(
-                (void *)&setup_chunk_data.m_chunk_color_data, sizeof(DirectX::XMFLOAT3), 1u);
-        }
-
-        setup_chunk_data.m_chunk.m_chunk_index = index;
-        return setup_chunk_data;
-    }
-
-  public:
-    void create_chunk(Renderer &renderer, const size_t index)
-    {
-        if (m_loaded_chunks.contains(index) || m_unloaded_chunks.contains(index))
-        {
-            return;
-        }
-
-        m_setup_chunk_futures_stack.emplace(
-            std::pair{renderer.m_copy_queue.m_monotonic_fence_value + 1,
-                      std::async(std::launch::async, &ChunkManager::_create_chunk, this, std::ref(renderer), index)});
-    }
-
-    void move_to_loaded_chunks(const u64 current_copy_queue_fence_value)
-    {
-        using namespace std::chrono_literals;
-
-        while (!m_setup_chunk_futures_stack.empty())
-        {
-            auto &setup_chunk_data = m_setup_chunk_futures_stack.front();
-            if (!setup_chunk_data.second.valid())
-            {
-                return;
-            }
-
-            switch (std::future_status status = setup_chunk_data.second.wait_for(0s); status)
-            {
-            case std::future_status::timeout: {
-
-                return;
-            }
-            break;
-
-            case std::future_status::ready: {
-                // If this condition is satisfied, the buffers are ready, so chunk is ready to be loaded :)
-                if (setup_chunk_data.first <= current_copy_queue_fence_value)
-                {
-                    SetupChunkData chunk_to_load = setup_chunk_data.second.get();
-
-                    const size_t num_vertices = chunk_to_load.m_chunk_position_data.size();
-                    const size_t chunk_index = chunk_to_load.m_chunk.m_chunk_index;
-
-                    m_loaded_chunks[chunk_index] = std::move(chunk_to_load.m_chunk);
-
-                    m_chunk_position_buffers[chunk_index] = std::move(chunk_to_load.m_chunk_position_buffer);
-
-                    m_chunk_color_buffers[chunk_index] = std::move(chunk_to_load.m_chunk_color_buffer);
-
-                    m_chunk_number_of_vertices[chunk_index] = num_vertices;
-
-                    m_setup_chunk_futures_stack.pop();
-                }
-                else
-                {
-                    return;
-                }
-            }
-            break;
-            }
-        }
-    }
-
-    static constexpr u32 NUMBER_OF_CHUNKS_PER_DIMENSION = 10;
-    static constexpr size_t NUMBER_OF_CHUNKS =
-        NUMBER_OF_CHUNKS_PER_DIMENSION * NUMBER_OF_CHUNKS_PER_DIMENSION * NUMBER_OF_CHUNKS_PER_DIMENSION;
-
-    std::unordered_map<size_t, Chunk> m_loaded_chunks{};
-    std::unordered_map<size_t, Chunk> m_unloaded_chunks{};
-
-    // NOTE : Chunks are considered to be setup when :
-    // (i) The result of async call (i.e the future) is ready,
-    // (ii) The fence value is < the current copy queue fence value.
-    // The stack consist of pairs of fence values , futures.
-    std::queue<std::pair<u64, std::future<SetupChunkData>>> m_setup_chunk_futures_stack{};
-
-    std::unordered_map<size_t, StructuredBuffer> m_chunk_position_buffers{};
-    std::unordered_map<size_t, StructuredBuffer> m_chunk_color_buffers{};
-    std::unordered_map<size_t, size_t> m_chunk_number_of_vertices{};
-};
 
 int main()
 {
@@ -524,8 +180,8 @@ int main()
     };
 
     // Execute and flush gpu so resources required for rendering (before the first frame) are ready.
-    renderer.execute_command_list(QueueType::Direct);
-    renderer.flush_gpu(QueueType::Direct);
+    renderer.m_direct_queue.execute_command_list();
+    renderer.m_direct_queue.flush_queue();
 
     Camera camera{};
     Timer timer{};
@@ -533,15 +189,18 @@ int main()
 
     u64 frame_count = 0;
 
-    // Test to check the async copy queue capabilities!
-    for (int i = 1; i < ChunkManager::NUMBER_OF_CHUNKS; i++)
-    {
-        chunk_manager.create_chunk(renderer, i);
-    }
-
     bool quit{false};
     while (!quit)
     {
+        if (frame_count == 0u)
+        {
+            // Test to check the async copy queue capabilities!
+            for (int i = 1; i < ChunkManager::NUMBER_OF_CHUNKS; i++)
+            {
+                chunk_manager.create_chunk(renderer, i);
+            }
+        }
+
         timer.start();
 
         MSG message = {};
@@ -568,14 +227,13 @@ int main()
 
         scene_buffer.update(&scene_buffer_data);
 
-        const auto &allocator = renderer.m_direct_queue.m_command_allocators[renderer.m_swapchain_backbuffer_index];
-        const auto &command_list = renderer.m_direct_queue.m_command_list;
-
         const auto &swapchain_index = renderer.m_swapchain_backbuffer_index;
 
         // Reset command allocator and command list.
-        throw_if_failed(allocator->Reset());
-        throw_if_failed(command_list->Reset(allocator.Get(), nullptr));
+        renderer.m_copy_queue.reset(swapchain_index);
+        renderer.m_direct_queue.reset(swapchain_index);
+
+        const auto &command_list = renderer.m_direct_queue.m_command_list;
 
         const auto &rtv_handle = renderer.m_swapchain_backbuffer_cpu_descriptor_handles[swapchain_index];
         const Microsoft::WRL::ComPtr<ID3D12Resource> swapchain_resource = renderer.m_resources[swapchain_index];
@@ -664,16 +322,20 @@ int main()
         command_list->ResourceBarrier(1u, &render_target_to_presentation_barrier);
 
         // Submit command list to queue for execution.
-        renderer.execute_command_list(QueueType::Direct);
+        renderer.m_direct_queue.execute_command_list();
+
+        renderer.m_copy_queue.execute_command_list();
 
         // Now, present the rendertarget and signal command queue.
         throw_if_failed(renderer.m_swapchain->Present(1u, 0u));
-        renderer.signal_fence(QueueType::Direct);
+        renderer.m_direct_queue.signal_fence(renderer.m_swapchain_backbuffer_index);
+        renderer.m_copy_queue.signal_fence(renderer.m_swapchain_backbuffer_index);
 
         renderer.m_swapchain_backbuffer_index = static_cast<u8>(renderer.m_swapchain->GetCurrentBackBufferIndex());
 
         // Wait for the previous frame (that is presenting to swpachain_backbuffer_index) to complete execution.
-        renderer.wait_for_fence_value_at_index(renderer.m_swapchain_backbuffer_index);
+        // NOTE : Do NOT do this for the copy queue if you want async copy.
+        renderer.m_direct_queue.wait_for_fence_value_at_index(renderer.m_swapchain_backbuffer_index);
 
         ++frame_count;
 
@@ -686,7 +348,8 @@ int main()
     ImGui_ImplWin32_Shutdown();
     ImGui::DestroyContext();
 
-    renderer.flush_gpu(QueueType::Direct);
+    renderer.m_direct_queue.flush_queue();
+    renderer.m_copy_queue.flush_queue();
 
     return 0;
 }
